refactor & improve perl language detection
